% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cs.shocks.R
\name{cs.shocks}
\alias{cs.shocks}
\title{Comparative Statics WRT Rho and Phi}
\usage{
cs.shocks(parameter = c("rho", "phi"), num, par.max, local = TRUE)
}
\arguments{
\item{parameter}{Pick either rho or phi to do comparative statics}

\item{num}{the Length of grid for rho or phi. For cluster use recommended to be a multiplier of 150}

\item{par.max}{The maximum of parameter value in the grid. The minimums are calculated automatically; for rho, it is negative of the rho max, for phi, it is the inverse of the max value. For sigma, the length of grid for values greater than one is equal to that of the ones less than one.}

\item{local}{Specifies the parallelization should happen on local machine or Acropolis cluster. Default is the local machine}
}
\description{
This function computes comparative statics wrt to rho (correlation coefficient of spouses' shocks)
and phi(the ratio of wife's standard deviation shock to her husband's).
}

