% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/E.1.u.d.der.R
\name{E.1.u.d.der}
\alias{E.1.u.d.der}
\title{Derivative of Expected Second Period Divorce Utility Given The First Period Info}
\usage{
E.1.u.d.der(S, type = c("u", "U"), spouse = c("h", "w"), i, r,
  sigma_eta_h = param()$sigma_eta_h, rho = param()$rho, phi = param()$phi)
}
\arguments{
\item{S}{Saving}

\item{type}{Could be individual utility (u) or the joint utility of the couple (U) in a unitary framework}

\item{spouse}{If the type is individual utility (u), we should specify we mean husband or wife utility}

\item{i}{The marriage index}

\item{r}{First period repetition. It is not necessary to be greater than one for the first period. It is needed for taking expectaions, which is required in the second period}

\item{sigma_eta_h}{The husband's variance of transitory shock. If not specified the default is the baseline value specified in the param()}

\item{rho}{The contemporaneous correlation coefficient of the husband and wife income shocks. If not specified the default is the baseline value specified in the param()}

\item{phi}{The ratio of the wife's standard deviation of the transitory shock to that of the husband. If not specified the default is the baseline value specified in the param()}
}
\description{
This function calculates the derivative of the expected second period divorce utility wrt to saving given the information revealed at the first period for a given level of savings for an individual family.
Essentially, it calculates the second period optimal consumption for the given level of saving in the state of divorce for different
relizations of the shocks in the second period and then substiute them in the individual utility functions (u) or the joint utility function(U).
Finally, it takes the average to return the expectations. To take the derivative, we passed it thuru expectations and then used the second period analytical results to take the derivative.
The second period optimal consumption analytical solutions are obtained with no optimization; just consumming the second period income.
}

