% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/E.1.u.m.lam.R
\name{E.1.u.m.lam}
\alias{E.1.u.m.lam}
\title{Expected Second Period Marriage Utility Given The First Period Info in the Collective Framework}
\usage{
E.1.u.m.lam(lambda, S, type = c("u", "U"), spouse = c("h", "w"), i, r)
}
\arguments{
\item{lambda}{Sharing rule}

\item{S}{Saving}

\item{type}{Could be individual utility (u) or the joint utility of the couple (U) in a unitary framework}

\item{spouse}{If the type is individual utility (u), we should specify we mean husband or wife utility}

\item{i}{The marriage index}

\item{r}{First period repetition. It is not necessary to be greater than one for the first period. It is needed for taking expectaions, which is required in the second period}
}
\description{
This function calculates the expected second period marriage utility given the information revealed at the first period for a given level of savings and a new sharing rule (for the second period) for an individual family.
Essentially, it calculates the second period optimal consumption for the given level of saving and updated sharing rule for different
relizations of the shocks in the second period and then substiute them in the individual utility functions (u) or the joint utility function(U).
Finally, it takes the average to return the expectations.
}

