% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/E.1.u.opt.R
\name{E.1.u.opt}
\alias{E.1.u.opt}
\title{Expected Value Function}
\usage{
E.1.u.opt(c2.opt, type = c("u", "U"), spouse = c("h", "w"))
}
\arguments{
\item{c2.opt}{The optimal consumption in the second}

\item{type}{Could be individual utility (u) or the joint utility of the couple (U) in a unitary framework}

\item{spouse}{If the type is individual utility (u), we should specify we mean husband or wife utility}
}
\description{
This function calculates the expected second period value function given the information revealed in the first period.
 Essentially, it evaluates the objective function in the optimal second period consumption values and takes the expectation.
}

