% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/feasible.lam.R
\name{feasible.lam}
\alias{feasible.lam}
\title{Find A Feasible Initial Point Where Both Participation Constraints Satisfy}
\usage{
feasible.lam(i, r, Extensive = FALSE)
}
\arguments{
\item{i}{The marriage index}

\item{r}{First period repetition. It is not necessary to be greater than one for the first period. It is needed for taking expectaions, which is required in the second period}

\item{Extensive}{Returns extra ouput including the whole feasible region, and unconstrained (with no change of sharing rule) optimum
consumption and saving.}
}
\description{
This function finds a pair of saving and sharing rule where satisfy the participation constraints of both couples under which
marriage can be a superior option over divorce. Given that we are assuming the sharing rule to be a half in the first period,
the optimal consumption in the first period is going to be equal. Therefore, given the value of saving, we can obtain the first period consumption levels.
These levels provide a feasible initial guess for the optimization routine. The loop of a simple two-dimensional grid search over saving and sharing rule will stop
immediately once we find a feasible pair.
In the following I describe how the procedure works. First we obtain the max and min of saving to pin down the two ends of the saving grid. Then,
find the optimal savings under divorce and marriage under the same sharing rule (unconstrained solution). If even with the same sharing rule the participation
contsraints of none of the spouses is binding, then the couple stays married with the old terms. Otherwise, we conduct a two-dimensional
search over saving and lambda that keeps both spouses better off in the marriage relative to divorce. We find the first of
these pairs as the initial guess for the optimization process. On the other hand, if the feasible set is empty the couple divorce
because there is no saving and sharing rule that dominates the expected utility of divorce for both spouses.
Notice if the income is missing the function return a vector of four NAs. This function can be used interchangably with feasible.region
}

