% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/incomeProcess.R
\name{incomeProcess}
\alias{incomeProcess}
\title{Simulate the Income Process}
\usage{
incomeProcess(sigma_eta_h, Rho, Phi)
}
\arguments{
\item{sigma_eta_h}{The husband's variance of transitory shock. If not specified the default is the baseline value specified in the param()}

\item{Rho}{The contemporaneous correlation coefficient of the husband and wife income shocks. If not specified the default is the baseline value specified in the param()}

\item{Phi}{The ratio of the wife's standard deviation of the transitory shock to that of the husband. If not specified the default is the baseline value specified in the param()}
}
\description{
This function simulates the income process for two spaouses for a given set of parameters of income shocks process. First, it checks the
covariance matrix determinant for the provided shock parameters. If the cov.det is negative, it exits and assigns FALSE to cov.det outcome.
Otherwise, it simulates the shocks. It is possible that the lifetime earnings of some individuals turn out to be negative, which is problematic computationally
and theoretecally. I have dropped those observations (in fact dropping the families for whom at least one spouse has negative lifetime earnings)
. However, this in fact messes the randomization. I assigend NA for both couples in both periods if the lifetime earninga of one of the
spouses is negative. In addition, we have identified those fammilies for who the total first period income is negative despite the fact that their lifetime earning is not.
}

