% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/obj.lam.R
\name{obj.lam}
\alias{obj.lam}
\title{Joint Objective Function Given Period One Info}
\usage{
obj.lam(c.h, c.w, S, lam, i, r)
}
\arguments{
\item{c.h}{First period husband consumption}

\item{c.w}{First period wife consumption}

\item{S}{Saving}

\item{lam}{Sharing rule for the second period}

\item{i}{The marriage index}

\item{r}{First period repetition. It is not necessary to be greater than one for the first period. It is needed for taking expectaions, which is required in the second period}
}
\description{
This function returns the (negative) value of joint objective function for a set of given first period varaibles as of period one information set.
}

