% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optimal.path.lam.R
\name{optimal.path.lam}
\alias{optimal.path.lam}
\title{Optimal Path in the Collective Framework}
\usage{
optimal.path.lam(Precise = FALSE, local = TRUE)
}
\arguments{
\item{Precise}{Use the optimization packages (nloptr) to derive first period optima. If it is off, which is the default, we use feasible.region to have a rough approximation of the optima.}

\item{local}{Specifies the parallelization should happen on local machine or Acropolis cluster. Default is the local machine}
}
\description{
This function returns the optimal path of consumption, saving, and sharing rule in each period and each marital states. It utilizes parallel computation. In the first period
we have four controls: husband's consumption, wife's consumption, saving, and the sharing rule next period provided staying married. The second period
has two controls: husband's and wife's consumptions. The last line of the code is to exit MPI when using SMOW package on cluster
}

