% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optimal.path.R
\name{optimal.path}
\alias{optimal.path}
\title{Initial parameters distribution}
\usage{
optimal.path(sigma_eta_h = param()$sigma_eta_h, rho = param()$rho,
  phi = param()$phi, Parallel = FALSE, local = TRUE)
}
\description{
This function simulate the income process for two spaouses
}

