% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/participate.lam.R
\name{participate.lam}
\alias{participate.lam}
\title{Participation Constraints}
\usage{
participate.lam(lam, S, i, r)
}
\arguments{
\item{lam}{Sharing rule}

\item{S}{Saving}

\item{i}{The marriage index}

\item{r}{First period repetition. It is not necessary to be greater than one for the first period. It is needed for taking expectaions, which is required in the second period}
}
\description{
This function checks whether the participation constraints of a specific family satisfy for a given value of saving and sharing rule
}

