% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/period.1.m.lam.R
\name{period.1.m.lam}
\alias{period.1.m.lam}
\title{The First Period Problem in the Collective Framework}
\usage{
period.1.m.lam(i, r)
}
\arguments{
\item{i}{The marriage index}

\item{r}{First period repetition. It is not necessary to be greater than one for the first period. It is needed for taking expectaions, which is required in the second period}
}
\description{
This function find the optimal spouses' consumption levels, saving, and sharing rule  in the following period.
The decision to divorce shows itself in the expected next period utility in the objective function; we use E.1.u.m.lam. This is very slow. The choice of algorithm is key.
This is the only algorithm that works with non-linear inequality constraints and equality constraint. We use the feasible.region rough approximation as the initial guesss.
}

