% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/period.1.m.R
\name{period.1.m}
\alias{period.1.m}
\title{The First Period Problem Given Decision to Stay Married in the Second Period}
\usage{
period.1.m(i, r, sigma_eta_h = param()$sigma_eta_h, rho = param()$rho,
  phi = param()$phi)
}
\arguments{
\item{i}{The marriage index}

\item{r}{First period repetition. It is not necessary to be greater than one for the first period. It is needed for taking expectaions, which is required in the second period}

\item{sigma_eta_h}{The husband's variance of transitory shock. If not specified the default is the baseline value specified in the param()}

\item{rho}{The contemporaneous correlation coefficient of the husband and wife income shocks. If not specified the default is the baseline value specified in the param()}

\item{phi}{The ratio of the wife's standard deviation of the transitory shock to that of the husband. If not specified the default is the baseline value specified in the param()}
}
\description{
This function find the optimal spouses' consumption levels and saving given the fact that they know that they are going to stay married in the following period.
The decision to divorce shows itself in the expected next period utility in the objective function; we use E.1.u.m as opposed to E.1.u.d
}

