% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/period.2.d.R
\name{period.2.d}
\alias{period.2.d}
\title{The Second Period Problem Given Decision to Divorce}
\usage{
period.2.d(S, i, r1, r2, sigma_eta_h = param()$sigma_eta_h,
  rho = param()$rho, phi = param()$phi)
}
\arguments{
\item{i}{The marriage index}

\item{r1}{First period repetition. It is not necessary to be greater than one for the first period. It is needed for taking expectaions, which is required in the second period}

\item{r2}{Second period repetition. It is needed for taking expectaions, which is required in the second period}

\item{sigma_eta_h}{The husband's variance of transitory shock. If not specified the default is the baseline value specified in the param()}

\item{rho}{The contemporaneous correlation coefficient of the husband and wife income shocks. If not specified the default is the baseline value specified in the param()}

\item{phi}{The ratio of the wife's standard deviation of the transitory shock to that of the husband. If not specified the default is the baseline value specified in the param()}
}
\description{
This function find the optimal spouses' consumption levels in the second period given their decision to get divorced.
There is no optimization involved here; they consume all their income in the second period.
}

