% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/period.2.m.lam.R
\name{period.2.m.lam}
\alias{period.2.m.lam}
\title{The Second Period Problem in the Collective Framework}
\usage{
period.2.m.lam(lambda, S, i, r1, r2, Analytical = TRUE)
}
\arguments{
\item{lambda}{Sharing Rule}

\item{S}{Saving}

\item{i}{The marriage index}

\item{r1}{First period repetition. It is not necessary to be greater than one for the first period. It is needed for taking expectaions, which is required in the second period}

\item{r2}{Second period repetition. It is needed for taking expectaions, which is required in the second period}

\item{Analytical}{If it is on it uses the analytical solution instead of using the optimization procedure.}
}
\description{
This function finds the optimal spouses' consumption levels in the second period for agiven level of saving and sharing rule.
We can use the optimization procedure or just use the analytical solution. The default is to use the analytical solution
 and it is necessary to be worked in some other parts of the code.
}

