% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/period.2.m.R
\name{period.2.m}
\alias{period.2.m}
\title{The Second Period Problem Given Decision to Stay Married}
\usage{
period.2.m(S, i, r1, r2, Analytical = TRUE,
  sigma_eta_h = param()$sigma_eta_h, rho = param()$rho, phi = param()$phi)
}
\arguments{
\item{S}{Saving}

\item{i}{The marriage index}

\item{r1}{First period repetition. It is not necessary to be greater than one for the first period. It is needed for taking expectaions, which is required in the second period}

\item{r2}{Second period repetition. It is needed for taking expectaions, which is required in the second period}

\item{Analytical}{If it is on it uses the analytical solution instead of using the optimization procedure.}

\item{sigma_eta_h}{The husband's variance of transitory shock. If not specified the default is the baseline value specified in the param()}

\item{rho}{The contemporaneous correlation coefficient of the husband and wife income shocks. If not specified the default is the baseline value specified in the param()}

\item{phi}{The ratio of the wife's standard deviation of the transitory shock to that of the husband. If not specified the default is the baseline value specified in the param()}
}
\description{
This function finds the optimal spouses' consumption levels in the second period given their decision to stay married.
We can use the optimization procedure or just use the analytical solution. The default is to use the analytical solution
 and it is necessary to be worked in some other parts of the code.
}

