% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/U2.lam.R
\name{U2.lam}
\alias{U2.lam}
\title{Joint Utility in the Collective Framework}
\usage{
U2.lam(lambda, ch, cw)
}
\arguments{
\item{lambda}{Sharing rule}

\item{ch}{Husband consumption}

\item{cw}{Wife consumption}
}
\description{
This function returns the joint utility function for a given set of sharing rule, husband, and wife consumptions.
}

